refactor(forms): move rich-text detection to @memberjunction/global
- Promote detectRichTextFormat + RichTextFormat to @memberjunction/global
  (util.ts) so any consumer can use it; remove the base-forms-local copy
- Make scan length caller-configurable (maxScanLength param) and lower the
  default from 8000 to 500 chars — if no markdown/HTML signal appears early,
  assume plain and save CPU on slower machines
- Move the unit tests to MJGlobal (17 tests, incl. scan-length coverage)
- Move EditRichTextMode JSDoc onto the getter (was on the backing field)
- Document the HTML-mode XSS protection: Angular's [innerHTML] binding
  auto-sanitizes in HTML SecurityContext (no MJGlobal sanitizer exists; its
  EscapeHTML would neutralize all markup, defeating HTML rendering)
